requirements.txt
streamlit
openai-whisper
gtts
requests